requirements.txt
streamlit==1.38.0
pandas==2.2.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
pillow==10.4.0
google-genai==0.6.0
xlsxwriter==3.2.0
xlrd==2.0.1
urllib3==2.2.2
